Move week day label to its own file
So it can be used in other places
