+++ Sources/SATSCore/Components/DatePickerView/Subviews/DatePickerView-WeekdaysView.swift
@@ -11,21 +11,9 @@ extension DatePickerView {
         var body: some View {
             HStack(spacing: 0) {
                 ForEach(helper.getWeekdays(), id: \.self) { day in
-                        Text(weekdaySingleCharacter(date: day))
-                    .frame(maxWidth: .infinity)
+                    WeekDayLabel(name: helper.weekdayName(for: day))
                 }
             }
-            .padding(.vertical, .spacingXXS)
-            .satsFont(.small, weight: .emphasis)
-            .foregroundColor(.onSurfaceDisabled)
-        }
-
-        // Should be replaced when targeting iOS 15 with
-        // Text(date, format: .dateTime.weekday(.narrow))
-        private func weekdaySingleCharacter(date: Date) -> String {
-            let dateFormatter = DateFormatter()
-            dateFormatter.dateFormat = "E"
-            return dateFormatter.string(from: date).prefix(1).uppercased()
         }
     }
 }
